feat: added gaussian model test
